Restructure project.
