+++ src/music.js
@@ -3,7 +3,7 @@ Processes music commands. Constructs Songs and manages MusicPlayers.
 */
 'use strict';
 const config = require('./json/config.json');
-const tool = require('./tool.js');
+const tool = require('./util/tool.js');
 
 const Song = require('./obj/music/Song.js');
 const MusicPlayer = require('./obj/music/MusicPlayer.js');
